requirements.txt
flask==2.3.3
boto3==1.34.69
botocore==1.34.69
authlib==1.2.1
werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0
flask-cors==4.0.0